добавлена геолокация закончен фильтр начата отправка фотографий

+++ src/com/sonetica/topface/net/SettingsRequest.java
@@ -8,12 +8,12 @@ import android.content.Context;
 public class SettingsRequest extends ApiRequest {
   // Data
   private String service = "settings";
-  public String name;     // новое имя пользователя в UTF-8
-  public int age = -1;    // возраст пользователя. Минимальное значение - 12, максимальное - 99. Если указано меньше минимального или болше максимального значение кропятся по ОДЗ
-  public int sex = -1;    // новый пол пользователя
-  public int lat = -1;    // долгота местонахождения пользователя
-  public int lng = -1;    // широта местонахождения пользователя
-  public int cityid = -1; // идентификатор города пользователя
+  public String name;      // новое имя пользователя в UTF-8
+  public int age = -1;     // возраст пользователя. Минимальное значение - 12, максимальное - 99. Если указано меньше минимального или болше максимального значение кропятся по ОДЗ
+  public int sex = -1;     // новый пол пользователя
+  public double lat = -1;  // долгота местонахождения пользователя
+  public double lng = -1;  // широта местонахождения пользователя
+  public int cityid = -1;  // идентификатор города пользователя
   //---------------------------------------------------------------------------
   public SettingsRequest(Context context) {
     super(context);
